chore: lints galore
