+++ packages/studio-base/src/components/TimeBasedChart/downsample.test.ts
@@ -2,23 +2,27 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
+import { iterateNormal } from "@foxglove/studio-base/components/Chart/datasets";
 import { downsampleTimeseries, downsampleScatter } from "./downsample";
 
 describe("downsampleTimeseries", () => {
-  const bounds = { width: 100, height: 100, x: { min: 0, max: 100 }, y: { min: 0, max: 100 } };
+  const bounds = {
+    width: 100,
+    height: 100,
+    bounds: { x: { min: 0, max: 100 }, y: { min: 0, max: 100 } },
+  };
 
   it("merges nearby points", () => {
     const result = downsampleTimeseries(
-      {
-        data: [
-          { x: 0, y: 0, value: 0 },
-          { x: 10, y: 0, value: 0 },
-          { x: 20, y: 0, value: 0 },
-          { x: 20, y: 1, value: 1 },
-          { x: 20, y: 10, value: 10 },
-          { x: 20, y: 20, value: 20 },
-        ],
-      },
+      iterateNormal,
+      [
+        { x: 0, y: 0, value: 0 },
+        { x: 10, y: 0, value: 0 },
+        { x: 20, y: 0, value: 0 },
+        { x: 20, y: 1, value: 1 },
+        { x: 20, y: 10, value: 10 },
+        { x: 20, y: 20, value: 20 },
+      ],
       bounds,
     );
     expect(result).toEqual({
@@ -33,16 +37,15 @@ describe("downsampleTimeseries", () => {
 
   it("preserves distinctly labeled segments", () => {
     const result = downsampleTimeseries(
-      {
-        data: [
-          { x: 0, y: 0, value: 0, label: "1" },
-          { x: 10, y: 0, value: 0, label: "2" },
-          { x: 20, y: 0, value: 0 },
-          { x: 20, y: 1, value: 1 },
-          { x: 20, y: 10, value: 10 },
-          { x: 20, y: 20, value: 20 },
-        ],
-      },
+      iterateNormal,
+      [
+        { x: 0, y: 0, value: 0, label: "1" },
+        { x: 10, y: 0, value: 0, label: "2" },
+        { x: 20, y: 0, value: 0 },
+        { x: 20, y: 1, value: 1 },
+        { x: 20, y: 10, value: 10 },
+        { x: 20, y: 20, value: 20 },
+      ],
       bounds,
     );
     expect(result).toEqual({
@@ -57,14 +60,13 @@ describe("downsampleTimeseries", () => {
 
   it("should keep the min/max values within an interval", () => {
     const result = downsampleTimeseries(
-      {
-        data: [
-          { x: 0, y: 0, value: 0 },
-          { x: 0, y: 100, value: 100 },
-          { x: 0, y: -20, value: -20 },
-          { x: 0, y: 4, value: 4 },
-        ],
-      },
+      iterateNormal,
+      [
+        { x: 0, y: 0, value: 0 },
+        { x: 0, y: 100, value: 100 },
+        { x: 0, y: -20, value: -20 },
+        { x: 0, y: 4, value: 4 },
+      ],
       bounds,
     );
     expect(result).toEqual({
@@ -79,15 +81,14 @@ describe("downsampleTimeseries", () => {
 
   it("should keep entry/exit datum to an interval", () => {
     const result = downsampleTimeseries(
-      {
-        data: [
-          { x: 0, y: 0, value: 0 },
-          { x: 1, y: 0, value: 0 },
-          { x: 1, y: 100, value: 100 },
-          { x: 1, y: 4, value: 4 },
-          { x: 2, y: 5, value: 5 },
-        ],
-      },
+      iterateNormal,
+      [
+        { x: 0, y: 0, value: 0 },
+        { x: 1, y: 0, value: 0 },
+        { x: 1, y: 100, value: 100 },
+        { x: 1, y: 4, value: 4 },
+        { x: 2, y: 5, value: 5 },
+      ],
       bounds,
     );
     expect(result).toEqual({
@@ -103,18 +104,21 @@ describe("downsampleTimeseries", () => {
 });
 
 describe("downsampleScatter", () => {
-  const bounds = { width: 100, height: 100, x: { min: 0, max: 100 }, y: { min: 0, max: 100 } };
+  const bounds = {
+    width: 100,
+    height: 100,
+    bounds: { x: { min: 0, max: 100 }, y: { min: 0, max: 100 } },
+  };
 
   it("ignores out of bounds points", () => {
     const result = downsampleScatter(
-      {
-        data: [
-          { x: -1, y: 0, value: 0 },
-          { x: 0, y: -1, value: -1 },
-          { x: 200, y: 0, value: 0 },
-          { x: 0, y: 200, value: 200 },
-        ],
-      },
+      iterateNormal,
+      [
+        { x: -1, y: 0, value: 0 },
+        { x: 0, y: -1, value: -1 },
+        { x: 200, y: 0, value: 0 },
+        { x: 0, y: 200, value: 200 },
+      ],
       bounds,
     );
     expect(result).toEqual({
@@ -127,14 +131,13 @@ describe("downsampleScatter", () => {
 
   it("merges nearby points", () => {
     const result = downsampleScatter(
-      {
-        data: [
-          { x: 0, y: 0, value: 0 },
-          { x: 0, y: 0.4, value: 0.4 },
-          { x: 0, y: 0.8, value: 0.8 },
-          { x: 0, y: 1, value: 1 },
-        ],
-      },
+      iterateNormal,
+      [
+        { x: 0, y: 0, value: 0 },
+        { x: 0, y: 0.4, value: 0.4 },
+        { x: 0, y: 0.8, value: 0.8 },
+        { x: 0, y: 1, value: 1 },
+      ],
       bounds,
     );
     expect(result).toEqual({
