+++ packages/studio-base/src/panels/Plot/csv.test.ts
@@ -11,7 +11,14 @@ describe("csv", () => {
         [
           {
             label: "label",
-            data: [{ receiveTime: { sec: 0, nsec: 0 }, x: 0, y: 0, value: 0 }],
+            data: [
+              {
+                receiveTime: [{ sec: 0, nsec: 0 }],
+                x: new Float32Array([0]),
+                y: new Float32Array([0]),
+                value: [0],
+              },
+            ],
           },
         ],
         "timestamp",
@@ -28,11 +35,11 @@ describe("csv", () => {
             label: "label",
             data: [
               {
-                receiveTime: { sec: 0, nsec: 0 },
-                x: 0,
+                receiveTime: [{ sec: 0, nsec: 0 }],
+                x: new Float32Array([0]),
                 // eslint-disable-next-line @typescript-eslint/no-loss-of-precision
-                y: Number(9999999999999001),
-                value: 9999999999999001n,
+                y: new Float32Array([Number(9999999999999001)]),
+                value: [9999999999999001n],
               },
             ],
           },
